requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
aioredis==2.0.1
    # via fastapi-mail
aiosmtplib==1.1.6
    # via fastapi-mail
anyio==3.6.1
    # via
    #   httpcore
    #   starlette
async-timeout==4.0.2
    # via
    #   aioredis
    #   redis
blinker==1.5
    # via fastapi-mail
certifi==2022.6.15
    # via
    #   httpcore
    #   httpx
click==8.1.3
    # via uvicorn
deprecated==1.2.13
    # via redis
dnspython==2.2.1
    # via
    #   email-validator
    #   pymongo
email-validator==1.2.1
    # via fastapi-mail
fakeredis==1.8.1
    # via fastapi-mail
fastapi==0.79.0
    # via
    #   -r requirements.in
    #   fastapi-mail
fastapi-mail==1.0.9
    # via -r requirements.in
h11==0.12.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via fastapi-mail
idna==3.3
    # via
    #   anyio
    #   email-validator
    #   rfc3986
jinja2==3.1.2
    # via fastapi-mail
markupsafe==2.1.1
    # via jinja2
packaging==21.3
    # via redis
pydantic==1.9.1
    # via
    #   fastapi
    #   fastapi-mail
pymongo[srv]==4.2.0
    # via -r requirements.in
pyparsing==3.0.9
    # via packaging
python-decouple==3.6
    # via -r requirements.in
python-multipart==0.0.5
    # via fastapi-mail
redis==4.3.4
    # via fakeredis
rfc3986[idna2008]==1.5.0
    # via httpx
six==1.16.0
    # via
    #   fakeredis
    #   python-multipart
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sortedcontainers==2.4.0
    # via fakeredis
starlette==0.19.1
    # via fastapi
typing-extensions==4.3.0
    # via
    #   aioredis
    #   pydantic
uvicorn==0.18.2
    # via -r requirements.in
wrapt==1.14.1
    # via deprecated
